+++ app_plus/src/main/java/org/gkisalatiga/plus/services/DataUpdater.kt
@@ -35,7 +35,7 @@ class DataUpdater(private val ctx: Context) {
 
     companion object {
         // The location of the remote feeds.json file to check for updates.
-        private const val FEEDS_JSON_SOURCE = "https://raw.githubusercontent.com/gkisalatiga/gkisplus-data-json/main/v2.1/data/feeds.min.json"
+        private const val FEEDS_JSON_SOURCE = "https://raw.githubusercontent.com/gkisalatiga/gkisplus-data-json/main/v2/data/feeds.min.json"
     }
 
     private fun getLastGalleryUpdate() : Int {